+++ ContentPatcher/Framework/Patches/EditDataPatchField.cs
@@ -27,10 +27,10 @@ namespace ContentPatcher.Framework.Patches
         public TokenisableJToken Value { get; }
 
         /// <summary>Whether the instance may change depending on the context.</summary>
-        public bool IsMutable => this.EntryKey.IsMutable || (this.Value?.IsMutable == true);
+        public bool IsMutable => this.Contextuals.IsMutable;
 
         /// <summary>Whether the instance is valid for the current context.</summary>
-        public bool IsReady => this.EntryKey.IsReady && this.Value?.IsReady != false;
+        public bool IsReady => this.Contextuals.IsReady;
 
 
         /*********